pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pathpal-api"
version = "0.1.0"
description = "PathPal API - Social Safety Navigation App"
authors = [
    {name = "PathPal Team", email = "something@domain.com"},
]
dependencies = [
    "fastapi[all]>=0.104.0",
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.3",
    "uvicorn[standard]>=0.24.0",
    "aiosqlite>=0.21.0",
    "greenlet>=3.2.4",
    "polyline>=2.0.2",
    "httpx>=0.25.2",
    "openai>=1.12.0",
    "fastapi-mail>=1.4.1",
    "aiofiles>=23.2.1",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
